feat: move versioning to collection level

+++ src/api/vectordb/versions/dtos.rs
@@ -4,7 +4,7 @@ use serde::Serialize;
 #[derive(Serialize)]
 pub struct VersionMetadata {
     pub hash: Hash,
-    pub version_number: u32,
+    pub version_number: u16,
     pub timestamp: u32,
     pub vector_count: u64,
 }
@@ -18,7 +18,7 @@ pub struct VersionListResponse {
 #[derive(Serialize)]
 pub struct CurrentVersionResponse {
     pub hash: Hash,
-    pub version_number: u32,
+    pub version_number: u16,
     pub timestamp: u64,
     pub vector_count: u64,
 }
